--- src/main/java/kitchenpos/common/kichenriders_client/DefaultKitchenridersClient.java
@@ -1,13 +1,0 @@
-package kitchenpos.common.kichenriders_client;
-
-import org.springframework.stereotype.Component;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-@Component
-public class DefaultKitchenridersClient implements KitchenridersClient {
-    @Override
-    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
-    }
-}
