--- src/main/java/kitchenpos/common/kichenriders_client/KitchenridersClient.java
@@ -1,8 +1,0 @@
-package kitchenpos.common.kichenriders_client;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public interface KitchenridersClient {
-    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
-}
