+++ src/main/java/kitchenpos/delivery_order/infra/kichenriders_client/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.delivery_order.infra.kichenriders_client;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
