+++ src/main/java/kitchenpos/eat_in_order/domain/OrderType.java
@@ -1,0 +1,5 @@
+package kitchenpos.eat_in_order.domain;
+
+public enum OrderType {
+    DELIVERY, TAKEOUT, EAT_IN
+}
