+++ src/main/java/kitchenpos/eat_in_order/domain/repository/JpaOrderTableRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.eat_in_order.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+import kitchenpos.eat_in_order.domain.OrderTable;
+
+public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
+}
